requirements.txt
shillelagh
appdirs==1.4.4
apsw==3.38.5.post1
astroid==2.11.7
attrs==21.4.0
beautifulsoup4==4.11.1
boto3==1.24.35
botocore==1.27.35
build==0.8.0
cachetools==5.2.0
certifi==2022.6.15
cfgv==3.3.1
charset-normalizer==2.1.0
click==8.1.3
codespell==2.1.0
coverage==6.4.2
dill==0.3.5.1
distlib==0.3.5
exceptiongroup==1.0.4
filelock==3.7.1
freezegun==1.2.1
google-auth==2.9.1
greenlet==1.1.2
html5lib==1.1
identify==2.5.2
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
jmespath==1.0.1
jsonpath-python==1.0.6
lazy-object-proxy==1.7.1
mccabe==0.7.0
multidict==6.0.2
nodeenv==1.7.0
numpy==1.23.1
packaging==21.3
pandas==1.4.3
pep517==0.12.0
pip-tools==6.8.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
prompt-toolkit==3.0.30
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyfakefs==4.6.3
Pygments==2.12.0
pylint==2.14.5
pyparsing==3.0.9
pytest==7.2.0
pytest-cov==3.0.0
pytest-integration==0.2.2
pytest-mock==3.8.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-cache==0.7.1
requests-mock==1.9.3
rsa==4.9
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
tabulate==0.8.9
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
typing_extensions==4.3.0
url-normalize==1.4.3
urllib3==1.26.10
virtualenv==20.15.1
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
yarl==1.8.1
